UI/Refactor: Replace web API drag and drop with Dart library
It turns out web API drag and drop does not support touch events!  Who
knew?  There is a web API for pointer events, which supports both touch
and mouse, but it's low-level and I'd have to build up a drag and drop
API on top of that.  So instead I went with an existing Dart library
that does custom drag and drop for touch and mouse together.