Convert drag and drog to TypeScript with interact.js
This conversion purposefully keeps things as close to the same format as
the prior drag and drop setup with minimal refactoring.  Currently the
`nt-drag-over` isn't working, but since we can refactor a lot of the
drag over code away it's not worth it to fix it as part of this work.